rename and move user abc-repo
