--- src/apps/user/repositories/__init__.py
@@ -1,3 +1,0 @@
-from .user_repository import IUserRepository
-
-__all__ = ['IUserRepository']
