+++ src/apps/user/services/authenticate_user_service.py
@@ -3,7 +3,7 @@ from passlib.context import CryptContext
 from src.apps.user.domain.entity import User
 from src.apps.user.dtos import AuthDTO
 from src.apps.user.exceptions import NotActivationExpire, UserPasswordException
-from src.apps.user.repositories import IUserRepository
+from src.apps.user.repository import IUserRepository
 from src.apps.user.services import GetUserService
