+++ src/apps/user/services/password_reset_user_service.py
@@ -1,7 +1,7 @@
 from src.apps.user.domain.entity import User
 from src.apps.user.dtos import AdminPasswordUpdateDTO, UserPasswordUpdateDTO
 from src.apps.user.exceptions import UserNotFoundByEmailException, UserPasswordException
-from src.apps.user.repositories import IUserRepository
+from src.apps.user.repository import IUserRepository
 from src.apps.user.services import AuthenticateUserService, CreateUserService
