+++ src/apps/user/services/update_user_service.py
@@ -2,7 +2,7 @@ from src.apps.user.domain.entity import User
 from src.apps.user.dtos import UserUpdateDTO
 from src.apps.user.exceptions import UserAlreadyExistsError
 from src.apps.user.protocols import JWTServiceProtocol
-from src.apps.user.repositories import IUserRepository
+from src.apps.user.repository import IUserRepository
 
 
 class UpdateUserService:
