+++ src/data_access/repositories/user_repository.py
@@ -5,7 +5,7 @@ from sqlalchemy import select, update
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from src.apps.user.domain.entity import User
-from src.apps.user.repositories import IUserRepository
+from src.apps.user.repository import IUserRepository
 from src.data_access.mappers.user_mapper import UserMapper
 from src.data_access.models import UserModel
